requirements.txt
Pillow==9.2.0
torch==1.12.1
torchvision==0.13.1
albumentations==1.3.0
numpy==1.21.6
pydicom==2.3.0
scikit-image==0.19.3

